Get winter temperatures and create utils.js file
